IBX-538: Renamed EzPublishCoreCollector to IbexaCoreCollector
