+++ src/bundle/Debug/Collector/IbexaCoreCollector.php
